Moved form-related modules into DataRepo/forms/, including the widgets directory.
Files checked in: - DataRepo/forms.py DataRepo/multiforms.py DataRepo/pager.py DataRepo/tests/views/search/test_download.py DataRepo/views/search/advanced.py DataRepo/views/search/basic.py DataRepo/views/search/download.py DataRepo/views/upload/submission.py DataRepo/widgets/__init__.py DataRepo/widgets/base.py DataRepo/widgets/search.py DataRepo/forms/__init__.py DataRepo/forms/forms.py DataRepo/forms/multiforms.py DataRepo/forms/widgets/__init__.py DataRepo/forms/widgets/base.py DataRepo/forms/widgets/pager.py DataRepo/forms/widgets/search.py
