--- -
@@ -1,2 +1,0 @@
-Name,Description
-Small OBOB,ob/ob and wildtype littermates were fasted 7 hours and infused with tracers
