+++ DataRepo/forms/forms.py
@@ -28,8 +28,8 @@ from DataRepo.utils.file_utils import (
     ensure_temporary_uploaded_file,
     is_excel,
 )
-from DataRepo.widgets.base import AutoCompleteTextInput, MultipleFileInput
-from DataRepo.widgets.search import RowsPerPageSelectWidget
+from DataRepo.forms.widgets.base import AutoCompleteTextInput, MultipleFileInput
+from DataRepo.forms.widgets.search import RowsPerPageSelectWidget
 
 # IMPORTANT NOTE ABOUT THE pos & posprefix FIELDS IN EACH AdvSearch FORM CLASSES:
