+++ DataRepo/tests/views/search/test_download.py
@@ -5,7 +5,7 @@ from io import BytesIO
 from django.core.management import call_command
 from django.http import StreamingHttpResponse
 
-from DataRepo.forms import AdvSearchDownloadForm
+from DataRepo.forms.forms import AdvSearchDownloadForm
 from DataRepo.loaders.study_loader import StudyV3Loader
 from DataRepo.models.peak_group import PeakGroup
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
