+++ DataRepo/views/search/basic.py
@@ -10,8 +10,8 @@ from django.http import Http404
 from django.shortcuts import render
 
 from DataRepo.formats.search_group import SearchGroup
-from DataRepo.forms import AdvSearchForm, AdvSearchPageForm
-from DataRepo.pager import Pager
+from DataRepo.forms.forms import AdvSearchForm, AdvSearchPageForm
+from DataRepo.forms.widgets.pager import Pager
 from DataRepo.views.search.advanced import AdvancedSearchView
 from DataRepo.views.utils import get_cookie
